Fix libgphoto2_port compilation warnings.
* gphoto2-port-info-list.c:
  * gp_port_info_list_append(...): Variables 'generic' and 'i' are changed to unsigned types as they are used in trivial 'for' loops already in unsigned comparison mode.
  * gp_port_info_list_get_info(...): Cast variable 'n' to unsigned in two places to avoid signed-unsigned compare warnings. This does not change behavior.
* gphoto2-port-log.c:
  * gp_log_remove_func(...): Change variable 'i' to unsigned type as it is used in trivial 'for' loop in unsigned comparison mode. Also casts variable 'id' to unsigned to avoid signed-unsigned comparison. This does not change code behavior.

+++ libgphoto2_port/libgphoto2_port/gphoto2-port-log.c
@@ -137,10 +137,10 @@ gpi_vsnprintf (const char* format, va_list args)
 int
 gp_log_remove_func (int id)
 {
-	int i;
+	unsigned int i;
 
 	for (i=0;i<log_funcs_count;i++) {
-		if (log_funcs[i].id == id) {
+		if (log_funcs[i].id == (unsigned int)id) {
 			memmove (log_funcs + i, log_funcs + i + 1, sizeof(LogFunc) * (log_funcs_count - i - 1));
 			log_funcs_count--;
 			return GP_OK;
